+++ src/main/java/basicstats/AddNumView.java
@@ -26,14 +26,14 @@ public class AddNumView implements View {
           public void actionPerformed(ActionEvent e) {
 
             for (Controller ctrl : controllers) {
-              ctrl.updateModels(Integer.parseInt(jtfNumber.getText()));
+              ctrl.updateModels(Double.parseDouble(jtfNumber.getText()));
             }
           }
         });
   }
 
   @Override
-  public void update(ArrayList<Integer> numbers) {
+  public void update(ArrayList<Double> numbers) {
     jtfNumber.setText("");
   }
