+++ src/main/java/basicstats/Controller.java
@@ -4,5 +4,5 @@ public interface Controller {
 
   public void addModel(Model model);
 
-  public void updateModels(Integer num);
+  public void updateModels(Double num);
 }
