+++ src/main/java/basicstats/View.java
@@ -6,7 +6,7 @@ import javax.swing.text.JTextComponent;
 // Interface for use of template method.
 public interface View {
 
-  public void update(ArrayList<Integer> numbers);
+  public void update(ArrayList<Double> numbers);
 
   public void reset();
